# citadel/server/openapi.yaml
openapi: 3.0.3
info:
  title: CITADEL Deep-Scan API
  version: "1.0"
  description: >-
    Read-only security & compliance review API. Uploaded code is only ever READ
    by scanners — never executed. All findings normalize to one schema and map to
    22 compliance frameworks. Most routes are permission-gated when the global
    `enforce` toggle is on; when off (default) the scan routes are open.
servers:
  - url: /
    description: Same-origin (the SPA is served by this API)
security:
  - bearerAuth: []
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Sign in (returns an access token + sets an httpOnly refresh cookie)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        "200":
          description: Authenticated, or an MFA challenge
          headers:
            Set-Cookie:
              description: "citadel_rt=<jwt>; HttpOnly; Secure; SameSite=Strict; Path=/api/auth"
              schema: { type: string }
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/TokenResponse"
                  - type: object
                    properties:
                      mfaRequired: { type: boolean }
                      mfaToken: { type: string }
        "401": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
...
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Exchange the httpOnly refresh cookie (or a body token) for a new access token
      security: []
      responses:
        "200": { description: New access token, content: { application/json: { schema: { type: object, properties: { token: { type: string }, expiresIn: { type: integer } } } } } }
        "401": { $ref: "#/components/responses/Error" }
  /api/auth/me:
    get:
      tags: [Auth]
      summary: The current authenticated user (no secrets)
      responses:
        "200": { description: User, content: { application/json: { schema: { $ref: "#/components/schemas/User" } } } }
        "401": { $ref: "#/components/responses/Error" }
...
  /api/auth/mfa/disable:
    post:
      tags: [Auth]
      summary: Disable MFA (requires the current password)
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [password], properties: { password: { type: string } } } } } }
      responses: { "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Ok" } } } } }
...
  /api/auth/oidc/callback:
    get:
      tags: [Auth]
      summary: OIDC redirect URI — verifies the id_token and hands an access token to the SPA
      security: []
      parameters:
        - { name: code, in: query, schema: { type: string } }
        - { name: state, in: query, schema: { type: string } }
      responses: { "200": { description: HTML token-handoff page (sets the refresh cookie) }, "400": { $ref: "#/components/responses/Error" }, "403": { $ref: "#/components/responses/Error" } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  responses:
    Error:
      description: Error (generic message; no stack/internal paths)
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
    Ok: { type: object, properties: { ok: { type: boolean } } }
    Error: { type: object, properties: { error: { type: string } } }
...
    TokenResponse:
      type: object
      properties:
        token: { type: string, description: Short-lived access JWT (Bearer) }
        expiresIn: { type: integer }
        user: { $ref: "#/components/schemas/User" }
    User:
      type: object
      description: Never includes pass/salt/mfaSecret.
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        role: { type: string, enum: [admin, analyst, auditor, viewer] }
        active: { type: boolean }
        permissions: { type: object, additionalProperties: { type: boolean } }
        mustChange: { type: boolean }
        mfaEnabled: { type: boolean }
...